+++ src/services/Organizations.php
@@ -8,19 +8,23 @@
 
 namespace flipbox\organizations\services;
 
+use craft\elements\db\UserQuery;
 use craft\helpers\ArrayHelper;
 use flipbox\ember\exceptions\RecordNotFoundException;
 use flipbox\ember\helpers\SiteHelper;
 use flipbox\ember\services\traits\elements\MultiSiteAccessor;
 use flipbox\organizations\db\OrganizationQuery;
 use flipbox\organizations\elements\Organization as OrganizationElement;
+use flipbox\organizations\Organizations as OrganizationPlugin;
+use craft\elements\User as UserElement;
+use flipbox\organizations\records\UserAssociation;
 use yii\base\Component;
 
 /**
  * @author Flipbox Factory <hello@flipboxfactory.com>
  * @since 1.0.0
  *
- * @method OrganizationElement create($config = [], string $toScenario = null)
+ * @method OrganizationElement create($config = [])
  * @method OrganizationElement find($identifier, int $siteId = null)
  * @method OrganizationElement get($identifier, int $siteId = null)
  * @method OrganizationQuery getQuery($criteria = [])
@@ -79,6 +83,152 @@ class Organizations extends Component
         }
 
         return $this->create($organization);
+    }
+
+    /**
+     * @param UserQuery $query
+     * @param OrganizationElement $organization
+     * @return bool
+     * @throws \Exception
+     */
+    public function saveAssociations(
+        UserQuery $query,
+        OrganizationElement $organization
+    ): bool {
+        /** @var UserElement[] $models */
+        if (null === ($models = $query->getCachedResult())) {
+            return true;
+        }
+
+        $models = $query->all();
+
+        $associationService = OrganizationPlugin::getInstance()->getOrganizationUserAssociations();
+
+        $query = $associationService->getQuery([
+            $associationService::SOURCE_ATTRIBUTE => $organization->getId() ?: false
+        ]);
+
+        $query->setCachedResult(
+            $this->toAssociations($models, $organization->getId())
+        );
+
+        return $associationService->save($query);
+    }
+
+    /**
+     * @param UserQuery $query
+     * @param OrganizationElement $organization
+     * @return bool
+     * @throws \Exception
+     */
+    public function dissociate(
+        UserQuery $query,
+        OrganizationElement $organization
+    ): bool {
+        return $this->associations(
+            $query,
+            $organization,
+            [
+                OrganizationPlugin::getInstance()->getOrganizationUserAssociations(),
+                'dissociate'
+            ]
+        );
+    }
+
+    /**
+     * @param UserQuery $query
+     * @param OrganizationElement $organization
+     * @return bool
+     * @throws \Exception
+     */
+    public function associate(
+        UserQuery $query,
+        OrganizationElement $organization
+    ): bool {
+        return $this->associations(
+            $query,
+            $organization,
+            [
+                OrganizationPlugin::getInstance()->getOrganizationUserAssociations(),
+                'associate'
+            ]
+        );
+    }
+
+    /**
+     * @param UserQuery $query
+     * @param OrganizationElement $organization
+     * @param callable $callable
+     * @return bool
+     */
+    protected function associations(UserQuery $query, OrganizationElement $organization, callable $callable)
+    {
+        /** @var UserElement[] $models */
+        if (null === ($models = $query->getCachedResult())) {
+            return true;
+        }
+
+        $models = ArrayHelper::index($models, 'id');
+
+        $success = true;
+        $ids = [];
+        $count = count($models);
+        $i = 0;
+        foreach ($this->toAssociations($models, $organization->getId()) as $association) {
+            if (true === call_user_func_array($callable, [$association, ++$i === $count])) {
+                ArrayHelper::remove($models, $association->userId);
+                $ids[] = $association->userId;
+                continue;
+            }
+
+            $success = false;
+        }
+
+        $query->id($ids);
+
+        if ($success === false) {
+            $query->setCachedResult($models);
+        }
+
+        return $success;
+    }
+
+    /**
+     * @param UserElement[] $users
+     * @param int $organizationId
+     * @return UserAssociation[]
+     */
+    private function toAssociations(
+        array $users,
+        int $organizationId
+    ) {
+        $associations = [];
+        $sortOrder = 1;
+
+        $associationService = OrganizationPlugin::getInstance()->getOrganizationUserAssociations();
+        $sortOrderAttribute = $associationService::SORT_ORDER_ATTRIBUTE;
+
+        $existingAssociations = $associationService->findAllByCriteria([
+            'where' => [
+                $associationService::TARGET_ATTRIBUTE => ArrayHelper::getColumn($users, 'id'),
+                $associationService::SOURCE_ATTRIBUTE => $organizationId,
+            ],
+            'indexBy' => $associationService::TARGET_ATTRIBUTE
+        ]);
+
+        foreach ($users as $user) {
+            if (null === ($association = ArrayHelper::remove($existingAssociations, $user->getId()))) {
+                $association = $associationService->create([
+                    'userId' => (int) $user->getId(),
+                    'organizationId' => (int) $organizationId
+                ]);
+            }
+
+            $association->{$sortOrderAttribute} = $sortOrder++;
+            $associations[] = $association;
+        }
+
+        return $associations;
     }
 
     /*******************************************